* Changes from code review; lots of cleanup, few bugfixes.
nettrace2ds.pm: Make regexes consistent with specified naming
  convention.  Merge three parallel loops into one loop.  Fix
  indentation style.  Merge parallel if branches to just using a
  variable.  Remove inaccurate comments.

DataSeriesFile.C: remove commented out hck, put in TODO explaining
  what needs to be fixed.

Extent.C: switch unsigned to uint32_t, remove cast leading to
  signed/unsigned warning.

ExtentSeries.C: Switch an AssertAlways to INVARIANT to eliminate size
  of variable problems.

MinMaxIndexModule.C: remove unused variable.

nettrace2ds.C: Explain the relationship between an info and convert
  pass; add TODO about switching to INVARIANT.  Add unknown type to
  FileT enum so it has a sane default. Switch all occurances of u_int
  to uint.  PCAPReader: eliminate need for buffer size constant,
  allocate to be the size specified in the pcap header, save said
  header so we can check everything will fit.  Expand inochange_time
  to inode_change_time in dataseries specification.  Later: if (false)
  a few debugging printfs.  NFSV3ReadDirPlusReplyHandler: split the big
  function into three pieces.  This made it clear there were a few
  missing datasize checks, add those in.  Fix a bug in use of strings.
  Improve a few comments to make it more clear what's being skiped.
  Modify the code so that all assertions are checked in Info mode.  Add
  a FATAL_ERROR to the getfattroffset function, I don't think it's
  correct, but I'm not sure what is correct.  fillcommonNFSRecord:
  move a bunch of assertions out so they run in info mode.  Fix
  comment around high bit being set.  main: cleanup argument handling 
  somewhat, merging common cases, again exposing a bug in one of the
  assertions.

sourcebyrange.C: remove /* leading to /* in /* warning.

+++ src/base/Extent.C
@@ -704,7 +704,7 @@ Extent::unpackData(const ExtentType *_type,
     byte compressed_variable_mode = from[6*4+1];
     byte type_name_len = from[6*4+2];
     
-    unsigned header_len = 6*4+4+type_name_len;
+    uint32_t header_len = 6*4+4+type_name_len;
     header_len += (4 - (header_len % 4))%4;
     AssertAlways(from.size() >= header_len,("Invalid extent data, too small"));
 
@@ -716,7 +716,7 @@ Extent::unpackData(const ExtentType *_type,
     int32 rounded_variable = compressed_variable_size;
     rounded_variable += (4-(rounded_variable%4))%4;
 
-    AssertAlways(header_len + rounded_fixed + rounded_variable == (int32)from.size(),
+    AssertAlways(header_len + rounded_fixed + rounded_variable == from.size(),
 		 ("Invalid extent data\n"));
 
     fixeddata.resize(nrecords * type->fixed_record_size);
